cast time_sec to long (Comment: #12344)
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/work/src@40596 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ uvm/impl/com/untangle/uvm/HostTableImpl.java
@@ -93,7 +93,7 @@ public class HostTableImpl implements HostTable
     {
         HostTableEntry entry = getHostTableEntry( address, true );
         long entryTime = System.currentTimeMillis();
-        long exitTime  = entryTime + (time_sec * 1000L);
+        long exitTime  = entryTime + (((long)time_sec) * 1000L);
 
         logger.info("Adding " + address.getHostAddress() + " to Penalty box for " + time_sec + " seconds");
